backend/requirements.txt
fastapi
uvicorn[standard]
aiomysql
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
PyPDF2
